backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
python-multipart==0.0.20
python-dotenv==1.0.1
openai==1.59.7
jinja2==3.1.5
aiosmtplib==3.0.2
email-validator==2.2.0
httpx==0.28.1
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
